Reset the branch for library restructuriong.
I accidentally deleted all the commit history on libstructure. This
meant that only the 8 new commits were in the history and all 110 or so
prior commits were deleted. I am doing a workaround by creating this new
branch and simply copying the files over to commit on top of the
original master branch.